refactor: rename DevServerOutputFinder to the more descriptive DevServerOutputTracker
